+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### May 2021 Release
 - Renamed all references to PSMA to Geoscape, reflecting the new data provider's name
 - The docker image build process has been completely redone
-- Changes in the Geoscape Admin Boundaries folder structure, resulting in duplicate Authority files, is now handled correctly. 
+- Changes in the Geoscape Admin Boundaries folder structure, resulting in duplicate Authority files, is now handled correctly.
+- Due to the above fix, all authority tables now have a standardised table structure of `gid, code, name, description`.
 ### August 2018 Release
 - GNAF now contains an *address_feature_pid* field for determining which addresses are new and which have changed. Previously the *address_detail_pid* on an address would change when a part of the address changed (e.g. 10 Smith St became 10 Smyth St) - even in cases where the building or dwelling it represents didn't change. This made determining which addresses were new and which had changed difficult. More info [here](https://www.psma.com.au/blog/blog/product-update-data-model-changes-improve-g-naf-product-scope)
 - *address_feature_pid* is in the `raw_gnaf_201808.address_feature` table. In the future, the data in this table will be added to the de-normalised schema (i.e. the gnaf_201808 schema).